[refactor] 푸시 알림 확인 후 앱 나갔다가 다시 들어갔을 경우 첫 화면 수정
